refactor: Sprint 6 hygiene — #111, #121, #122, #123
#111: Split 2095-line postgres.go into 7 focused files in the db package.
  postgres.go keeps struct, migrations, shared helpers (rowToMemory,
  rowToChunk, unwrapTx, execer). Method groups → postgres_memory.go,
  postgres_chunk.go, postgres_relationship.go, postgres_prune.go,
  postgres_episode.go, postgres_feedback.go. Pure file split, no logic
  changes; go build and go vet both exit 0.

#121: Fix byte-len truncation in SummarizeContent and ClaudeSummarize.
  len(content) was slicing multi-byte UTF-8 chars; now uses
  utf8.RuneCountInString and []rune slicing via truncateRunes helper.

#122: Remove stale Jaccard TODO in Consolidate — bigramJaccard is already
  implemented and used by ConsolidateWithClaude; comment now points there.

#123: Replace fmt.Sprintf string concat in ListMemories with clause-slice
  builder + strings.Join. Placeholder args still parameterized (no SQL
  injection risk), but code is now more readable and maintainable.
