+++ internal/db/postgres.go
@@ -2,7 +2,6 @@ package db
 
 import (
 	"context"
-	"crypto/sha256"
 	"embed"
 	"encoding/json"
 	"fmt"
@@ -285,15 +284,6 @@ type pgxTx struct{ tx pgx.Tx }
 func (t *pgxTx) Commit(ctx context.Context) error   { return t.tx.Commit(ctx) }
 func (t *pgxTx) Rollback(ctx context.Context) error { return t.tx.Rollback(ctx) }
 
-// Begin starts a new transaction.
-func (b *PostgresBackend) Begin(ctx context.Context) (Tx, error) {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return nil, err
-	}
-	return &pgxTx{tx: tx}, nil
-}
-
 // unwrapTx extracts the underlying pgx.Tx from a Tx interface value.
 func unwrapTx(t Tx) (pgx.Tx, error) {
 	pt, ok := t.(*pgxTx)
@@ -301,6 +291,11 @@ func unwrapTx(t Tx) (pgx.Tx, error) {
 		return nil, fmt.Errorf("unwrapTx: expected *pgxTx, got %T", t)
 	}
 	return pt.tx, nil
+}
+
+// execer is satisfied by both *pgxpool.Pool and pgx.Tx.
+type execer interface {
+	Exec(ctx context.Context, sql string, args ...any) (pgconn.CommandTag, error)
 }
 
 // ── Project metadata ──────────────────────────────────────────────────────────
@@ -340,1538 +335,6 @@ func (b *PostgresBackend) SetMetaTx(ctx context.Context, tx Tx, project, key, va
 		project, key, value,
 	)
 	return err
-}
-
-// ── Memory CRUD ───────────────────────────────────────────────────────────────
-
-func contentHash(content string) string {
-	h := sha256.Sum256([]byte(content))
-	return fmt.Sprintf("%x", h)
-}
-
-func (b *PostgresBackend) StoreMemory(ctx context.Context, m *types.Memory) error {
-	return b.storeMemoryExec(ctx, b.pool, m)
-}
-
-func (b *PostgresBackend) StoreMemoryTx(ctx context.Context, tx Tx, m *types.Memory) error {
-	raw, err := unwrapTx(tx)
-	if err != nil {
-		return err
-	}
-	return b.storeMemoryExec(ctx, raw, m)
-}
-
-// execer is satisfied by both *pgxpool.Pool and pgx.Tx.
-type execer interface {
-	Exec(ctx context.Context, sql string, args ...any) (pgconn.CommandTag, error)
-}
-
-func (b *PostgresBackend) storeMemoryExec(ctx context.Context, ex execer, m *types.Memory) error {
-	now := time.Now().UTC()
-	m.CreatedAt = now
-	m.UpdatedAt = now
-	m.LastAccessed = now
-	m.Project = b.project
-	hash := contentHash(m.Content)
-	m.ContentHash = &hash
-
-	tagsJSON, err := json.Marshal(m.Tags)
-	if err != nil {
-		return fmt.Errorf("marshal tags: %w", err)
-	}
-
-	episodeID := m.EpisodeID
-	var episodeArg any
-	if episodeID == "" {
-		episodeArg = nil
-	} else {
-		episodeArg = episodeID
-	}
-
-	// Seed dynamic_importance from static importance using Feature 2 formula.
-	if m.DynamicImportance == nil {
-		di := math.Max(0.1, (5.0-float64(m.Importance))/3.0)
-		m.DynamicImportance = &di
-	}
-
-	_, err = ex.Exec(ctx, `
-		INSERT INTO memories
-		  (id, content, memory_type, project, tags,
-		   importance, access_count, last_accessed, created_at, updated_at,
-		   immutable, expires_at, content_hash, storage_mode, episode_id,
-		   dynamic_importance)
-		VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16)`,
-		m.ID, m.Content, m.MemoryType, m.Project, tagsJSON,
-		m.Importance, m.AccessCount, now, now, now,
-		m.Immutable, m.ExpiresAt, hash, m.StorageMode, episodeArg,
-		m.DynamicImportance,
-	)
-	return err
-}
-
-func (b *PostgresBackend) GetMemory(ctx context.Context, id string) (*types.Memory, error) {
-	row, err := b.pool.Query(ctx,
-		"SELECT * FROM memories WHERE id=$1 AND project=$2 AND valid_to IS NULL", id, b.project)
-	if err != nil {
-		return nil, err
-	}
-	m, err := pgx.CollectOneRow(row, rowToMemory)
-	if err == pgx.ErrNoRows {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-	// Integrity check
-	if m.ContentHash != nil {
-		expected := contentHash(m.Content)
-		if *m.ContentHash != expected {
-			slog.Warn("INTEGRITY: content_hash mismatch",
-				"id", m.ID,
-				"stored", (*m.ContentHash)[:8],
-				"expected", expected[:8],
-			)
-		}
-	}
-	return m, nil
-}
-
-func (b *PostgresBackend) GetMemoriesByIDs(ctx context.Context, project string, ids []string) ([]*types.Memory, error) {
-	if len(ids) == 0 {
-		return nil, nil
-	}
-	rows, err := b.pool.Query(ctx,
-		"SELECT * FROM memories WHERE project=$1 AND id=ANY($2) AND valid_to IS NULL",
-		project, ids,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var mems []*types.Memory
-	for rows.Next() {
-		m, err := rowToMemory(rows)
-		if err != nil {
-			return nil, err
-		}
-		mems = append(mems, m)
-	}
-	return mems, rows.Err()
-}
-
-func (b *PostgresBackend) UpdateMemory(
-	ctx context.Context, id string,
-	content *string, tags []string, importance *int,
-) (*types.Memory, error) {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return nil, err
-	}
-	defer tx.Rollback(ctx) //nolint:errcheck
-
-	// Lock the row for the duration of the read-modify-write to prevent races.
-	row, err := tx.Query(ctx,
-		"SELECT * FROM memories WHERE id=$1 AND project=$2 AND valid_to IS NULL FOR UPDATE",
-		id, b.project)
-	if err != nil {
-		return nil, err
-	}
-	m, err := pgx.CollectOneRow(row, rowToMemory)
-	if err == pgx.ErrNoRows {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-	if m.Immutable {
-		return nil, fmt.Errorf("memory %q is immutable and cannot be updated", id)
-	}
-
-	// Snapshot current state into memory_versions before applying changes.
-	if err := b.versionMemoryTx(ctx, tx, m, types.VersionChangeUpdate, ""); err != nil {
-		return nil, fmt.Errorf("version snapshot: %w", err)
-	}
-
-	now := time.Now().UTC()
-	if content != nil {
-		m.Content = *content
-	}
-	if tags != nil {
-		m.Tags = tags
-	}
-	if importance != nil {
-		m.Importance = *importance
-	}
-
-	tagsJSON, err := json.Marshal(m.Tags)
-	if err != nil {
-		return nil, fmt.Errorf("marshal tags: %w", err)
-	}
-
-	if content != nil {
-		hash := contentHash(m.Content)
-		m.ContentHash = &hash
-		_, err = tx.Exec(ctx,
-			"UPDATE memories SET content=$1, tags=$2, importance=$3, updated_at=$4, content_hash=$5 WHERE id=$6 AND project=$7",
-			m.Content, tagsJSON, m.Importance, now, hash, id, b.project,
-		)
-	} else {
-		_, err = tx.Exec(ctx,
-			"UPDATE memories SET content=$1, tags=$2, importance=$3, updated_at=$4 WHERE id=$5 AND project=$6",
-			m.Content, tagsJSON, m.Importance, now, id, b.project,
-		)
-	}
-	if err != nil {
-		return nil, err
-	}
-	if err := tx.Commit(ctx); err != nil {
-		return nil, err
-	}
-	m.UpdatedAt = now
-	return m, nil
-}
-
-// DeleteMemory removes a memory and all its dependent data (chunks, relationships)
-// in a single atomic transaction. Routes through DeleteMemoryAtomic — keeping both
-// for interface compatibility.
-func (b *PostgresBackend) DeleteMemory(ctx context.Context, id string) (bool, error) {
-	return b.DeleteMemoryAtomic(ctx, b.project, id, false)
-}
-
-func (b *PostgresBackend) DeleteMemoryAtomic(ctx context.Context, project, id string, force bool) (bool, error) {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return false, err
-	}
-	defer tx.Rollback(ctx) //nolint:errcheck
-
-	var immutable bool
-	err = tx.QueryRow(ctx,
-		"SELECT immutable FROM memories WHERE id=$1 AND project=$2 FOR UPDATE",
-		id, project,
-	).Scan(&immutable)
-	if err == pgx.ErrNoRows {
-		return false, nil
-	}
-	if err != nil {
-		return false, err
-	}
-
-	if !force && immutable {
-		return false, fmt.Errorf("cannot delete immutable memory %s; use force=true only for rollback", id)
-	}
-	if force && immutable {
-		slog.Warn("force-deleting immutable memory (rollback path)", "id", id)
-	}
-
-	if _, err := tx.Exec(ctx, "DELETE FROM chunks WHERE memory_id=$1", id); err != nil {
-		return false, err
-	}
-	if _, err := tx.Exec(ctx, "DELETE FROM relationships WHERE source_id=$1 OR target_id=$1", id); err != nil {
-		return false, err
-	}
-	tag, err := tx.Exec(ctx, "DELETE FROM memories WHERE id=$1 AND project=$2", id, project)
-	if err != nil {
-		return false, err
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return false, err
-	}
-	return tag.RowsAffected() > 0, nil
-}
-
-// MergeMemoriesAtomic updates winnerID's content (if newContent is non-empty) and
-// deletes loserID in a single transaction (#104). This prevents the state where
-// winnerID has merged content but loserID was never deleted on a crash.
-func (b *PostgresBackend) MergeMemoriesAtomic(ctx context.Context, project, winnerID, loserID, newContent string) error {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback(ctx) //nolint:errcheck
-
-	if newContent != "" {
-		now := time.Now().UTC()
-		hash := contentHash(newContent)
-		if _, err := tx.Exec(ctx,
-			"UPDATE memories SET content=$1, content_hash=$2, updated_at=$3 WHERE id=$4 AND project=$5",
-			newContent, hash, now, winnerID, project,
-		); err != nil {
-			return fmt.Errorf("MergeMemoriesAtomic update winner: %w", err)
-		}
-	}
-
-	// Lock loser row to ensure it still exists before deleting.
-	var immutable bool
-	err = tx.QueryRow(ctx,
-		"SELECT immutable FROM memories WHERE id=$1 AND project=$2 FOR UPDATE",
-		loserID, project,
-	).Scan(&immutable)
-	if err == pgx.ErrNoRows {
-		// Loser already gone — treat as already merged.
-		return tx.Commit(ctx)
-	}
-	if err != nil {
-		return fmt.Errorf("MergeMemoriesAtomic lock loser: %w", err)
-	}
-	if immutable {
-		return fmt.Errorf("MergeMemoriesAtomic: loser %s is immutable", loserID)
-	}
-
-	if _, err := tx.Exec(ctx, "DELETE FROM chunks WHERE memory_id=$1", loserID); err != nil {
-		return fmt.Errorf("MergeMemoriesAtomic delete chunks: %w", err)
-	}
-	if _, err := tx.Exec(ctx, "DELETE FROM relationships WHERE source_id=$1 OR target_id=$1", loserID); err != nil {
-		return fmt.Errorf("MergeMemoriesAtomic delete relationships: %w", err)
-	}
-	if _, err := tx.Exec(ctx, "DELETE FROM memories WHERE id=$1 AND project=$2", loserID, project); err != nil {
-		return fmt.Errorf("MergeMemoriesAtomic delete loser: %w", err)
-	}
-
-	return tx.Commit(ctx)
-}
-
-func (b *PostgresBackend) ListMemories(ctx context.Context, project string, opts ListOptions) ([]*types.Memory, error) {
-	q := "SELECT * FROM memories WHERE project=$1 AND valid_to IS NULL"
-	args := []any{project}
-	n := 2
-
-	if opts.MemoryType != nil {
-		q += fmt.Sprintf(" AND memory_type=$%d", n)
-		args = append(args, *opts.MemoryType)
-		n++
-	}
-	if opts.ImportanceCeiling != nil {
-		q += fmt.Sprintf(" AND importance<=$%d", n)
-		args = append(args, *opts.ImportanceCeiling)
-		n++
-	}
-	for _, tag := range opts.Tags {
-		q += fmt.Sprintf(" AND tags @> $%d::jsonb", n)
-		j, _ := json.Marshal([]string{tag})
-		args = append(args, string(j))
-		n++
-	}
-
-	limit := opts.Limit
-	if limit <= 0 {
-		limit = 20
-	}
-	q += fmt.Sprintf(" ORDER BY updated_at DESC LIMIT $%d OFFSET $%d", n, n+1)
-	args = append(args, limit, opts.Offset)
-
-	rows, err := b.pool.Query(ctx, q, args...)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, rowToMemory)
-}
-
-func (b *PostgresBackend) TouchMemory(ctx context.Context, id string) error {
-	_, err := b.pool.Exec(ctx,
-		"UPDATE memories SET access_count=access_count+1, last_accessed=$1 WHERE id=$2",
-		time.Now().UTC(), id,
-	)
-	return err
-}
-
-// TouchMemories batch-updates access_count and last_accessed for multiple memories
-// in a single query (#117), replacing the N+1 TouchMemory calls in RecallWithOpts.
-func (b *PostgresBackend) TouchMemories(ctx context.Context, ids []string) error {
-	if len(ids) == 0 {
-		return nil
-	}
-	_, err := b.pool.Exec(ctx,
-		"UPDATE memories SET access_count=access_count+1, last_accessed=$1 WHERE id=ANY($2)",
-		time.Now().UTC(), ids,
-	)
-	return err
-}
-
-// ── Chunk CRUD ────────────────────────────────────────────────────────────────
-
-func (b *PostgresBackend) StoreChunks(ctx context.Context, chunks []*types.Chunk) error {
-	return b.storeChunksExec(ctx, b.pool, chunks)
-}
-
-func (b *PostgresBackend) StoreChunksTx(ctx context.Context, tx Tx, chunks []*types.Chunk) error {
-	raw, err := unwrapTx(tx)
-	if err != nil {
-		return err
-	}
-	return b.storeChunksExec(ctx, raw, chunks)
-}
-
-func (b *PostgresBackend) storeChunksExec(ctx context.Context, ex execer, chunks []*types.Chunk) error {
-	if len(chunks) == 0 {
-		return nil
-	}
-	const chunkSQL = `
-		INSERT INTO chunks (id, memory_id, project, chunk_text, chunk_index,
-		                    chunk_hash, embedding, section_heading, chunk_type)
-		VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9)
-		ON CONFLICT (id) DO NOTHING`
-	for _, c := range chunks {
-		var embParam any
-		if len(c.Embedding) > 0 {
-			embParam = pgvector.NewVector(c.Embedding)
-		}
-		_, err := ex.Exec(ctx, chunkSQL,
-			c.ID, c.MemoryID, c.Project,
-			c.ChunkText, c.ChunkIndex, c.ChunkHash,
-			embParam, c.SectionHeading, c.ChunkType,
-		)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// chunkCols is an explicit column list for chunks SELECTs. Using SELECT c.*
-// breaks positional scanning when migrations reorder columns (e.g. after
-// 003_pgvector.sql drops the old BYTEA embedding and adds vector(768)).
-const chunkCols = "c.id, c.memory_id, c.project, c.chunk_text, c.chunk_index, c.chunk_hash, c.embedding, c.section_heading, c.chunk_type, c.last_matched"
-
-func (b *PostgresBackend) GetChunksForMemory(ctx context.Context, memoryID string) ([]*types.Chunk, error) {
-	rows, err := b.pool.Query(ctx,
-		"SELECT "+chunkCols+" FROM chunks c WHERE c.memory_id=$1 ORDER BY c.chunk_index", memoryID,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, rowToChunk)
-}
-
-func (b *PostgresBackend) GetAllChunksWithEmbeddings(ctx context.Context, project string, limit int) ([]*types.Chunk, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT `+chunkCols+` FROM chunks c
-		JOIN memories m ON m.id = c.memory_id
-		WHERE c.embedding IS NOT NULL AND m.project=$1 AND m.valid_to IS NULL
-		ORDER BY m.last_accessed DESC
-		LIMIT $2`, project, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, rowToChunk)
-}
-
-func (b *PostgresBackend) GetAllChunkTexts(ctx context.Context, project string, limit int) ([]string, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT c.chunk_text FROM chunks c
-		JOIN memories m ON m.id = c.memory_id
-		WHERE m.project=$1 AND m.valid_to IS NULL LIMIT $2`, project, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var texts []string
-	for rows.Next() {
-		var t string
-		if err := rows.Scan(&t); err != nil {
-			return nil, err
-		}
-		texts = append(texts, t)
-	}
-	return texts, rows.Err()
-}
-
-func (b *PostgresBackend) GetChunksForMemories(ctx context.Context, memoryIDs []string) ([]*types.Chunk, error) {
-	if len(memoryIDs) == 0 {
-		return nil, nil
-	}
-	rows, err := b.pool.Query(ctx, `
-		SELECT `+chunkCols+` FROM chunks c
-		WHERE c.memory_id = ANY($1) AND c.embedding IS NOT NULL
-		ORDER BY c.chunk_index`, memoryIDs,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, rowToChunk)
-}
-
-func (b *PostgresBackend) ChunkHashExists(ctx context.Context, chunkHash, _ string) (bool, error) {
-	var exists bool
-	err := b.pool.QueryRow(ctx, `
-		SELECT EXISTS(
-			SELECT 1 FROM chunks c
-			JOIN memories m ON m.id = c.memory_id
-			WHERE c.chunk_hash=$1 AND m.project=$2
-		)`, chunkHash, b.project,
-	).Scan(&exists)
-	return exists, err
-}
-
-func (b *PostgresBackend) DeleteChunksForMemory(ctx context.Context, memoryID string) error {
-	_, err := b.pool.Exec(ctx, "DELETE FROM chunks WHERE memory_id=$1", memoryID)
-	return err
-}
-
-func (b *PostgresBackend) DeleteChunksForMemoryTx(ctx context.Context, tx Tx, memoryID string) error {
-	raw, err := unwrapTx(tx)
-	if err != nil {
-		return err
-	}
-	_, err = raw.Exec(ctx, "DELETE FROM chunks WHERE memory_id=$1", memoryID)
-	return err
-}
-
-func (b *PostgresBackend) DeleteChunksByIDs(ctx context.Context, chunkIDs []string) (int, error) {
-	if len(chunkIDs) == 0 {
-		return 0, nil
-	}
-	tag, err := b.pool.Exec(ctx, "DELETE FROM chunks WHERE id=ANY($1)", chunkIDs)
-	return int(tag.RowsAffected()), err
-}
-
-func (b *PostgresBackend) NullAllEmbeddings(ctx context.Context, project string) (int, error) {
-	tag, err := b.pool.Exec(ctx,
-		"UPDATE chunks SET embedding=NULL WHERE memory_id IN (SELECT id FROM memories WHERE project=$1)",
-		project,
-	)
-	return int(tag.RowsAffected()), err
-}
-
-func (b *PostgresBackend) NullAllEmbeddingsTx(ctx context.Context, tx Tx, project string) (int, error) {
-	raw, err := unwrapTx(tx)
-	if err != nil {
-		return 0, err
-	}
-	tag, err := raw.Exec(ctx,
-		"UPDATE chunks SET embedding=NULL WHERE memory_id IN (SELECT id FROM memories WHERE project=$1)",
-		project,
-	)
-	return int(tag.RowsAffected()), err
-}
-
-func (b *PostgresBackend) GetChunksPendingEmbedding(ctx context.Context, project string, limit int) ([]*types.Chunk, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT `+chunkCols+` FROM chunks c
-		JOIN memories m ON m.id = c.memory_id
-		WHERE m.project=$1 AND c.embedding IS NULL
-		ORDER BY m.last_accessed DESC
-		LIMIT $2`, project, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return pgx.CollectRows(rows, rowToChunk)
-}
-
-func (b *PostgresBackend) UpdateChunkEmbedding(ctx context.Context, chunkID string, embedding []float32) (int, error) {
-	tag, err := b.pool.Exec(ctx,
-		"UPDATE chunks SET embedding=$1 WHERE id=$2", pgvector.NewVector(embedding), chunkID,
-	)
-	return int(tag.RowsAffected()), err
-}
-
-func (b *PostgresBackend) VectorSearch(ctx context.Context, project string, queryVec []float32, limit int) ([]VectorHit, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT c.id, c.memory_id,
-		       c.embedding <=> $1::vector AS distance,
-		       c.chunk_text, c.chunk_index, c.section_heading
-		FROM chunks c
-		WHERE c.project = $2 AND c.embedding IS NOT NULL
-		ORDER BY c.embedding <=> $1::vector
-		LIMIT $3`,
-		pgvector.NewVector(queryVec), project, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var hits []VectorHit
-	for rows.Next() {
-		var h VectorHit
-		if err := rows.Scan(&h.ChunkID, &h.MemoryID, &h.Distance,
-			&h.ChunkText, &h.ChunkIndex, &h.SectionHeading); err != nil {
-			return nil, err
-		}
-		hits = append(hits, h)
-	}
-	return hits, rows.Err()
-}
-
-// ChunkEmbeddingDistance returns the minimum cosine distance between any chunk of
-// memAID and any chunk of memBID. Uses a LATERAL join so the HNSW index on
-// chunks.embedding is used for each probe — O(N·log M) instead of O(N×M) (#114).
-func (b *PostgresBackend) ChunkEmbeddingDistance(ctx context.Context, memAID, memBID string) (float64, error) {
-	var dist *float64
-	err := b.pool.QueryRow(ctx, `
-		SELECT MIN(sq.dist)
-		FROM chunks ca
-		JOIN LATERAL (
-			SELECT ca.embedding <=> cb.embedding AS dist
-			FROM chunks cb
-			WHERE cb.memory_id = $2 AND cb.embedding IS NOT NULL
-			ORDER BY dist ASC
-			LIMIT 1
-		) sq ON true
-		WHERE ca.memory_id = $1 AND ca.embedding IS NOT NULL`,
-		memAID, memBID,
-	).Scan(&dist)
-	if err != nil {
-		return 2.0, err
-	}
-	if dist == nil {
-		return 2.0, nil // no embedded chunks
-	}
-	return *dist, nil
-}
-
-func (b *PostgresBackend) UpdateChunkLastMatched(ctx context.Context, chunkID string) error {
-	_, err := b.pool.Exec(ctx,
-		"UPDATE chunks SET last_matched=NOW() WHERE id=$1", chunkID,
-	)
-	return err
-}
-
-func (b *PostgresBackend) GetPendingEmbeddingCount(ctx context.Context, project string) (int, error) {
-	var count int
-	err := b.pool.QueryRow(ctx, `
-		SELECT COUNT(*) FROM chunks c
-		JOIN memories m ON m.id = c.memory_id
-		WHERE m.project=$1 AND m.valid_to IS NULL AND c.embedding IS NULL`, project,
-	).Scan(&count)
-	return count, err
-}
-
-// ── Relationship CRUD ─────────────────────────────────────────────────────────
-
-// StoreRelationship upserts a directed relationship between two memories.
-// Both existence checks and the upsert run inside a single transaction (#110)
-// so a concurrent DeleteMemory between the check and the insert is prevented
-// by the SELECT ... FOR UPDATE row locks.
-func (b *PostgresBackend) StoreRelationship(ctx context.Context, rel *types.Relationship) error {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback(ctx)
-
-	var dummy int
-	if err := tx.QueryRow(ctx,
-		"SELECT 1 FROM memories WHERE id=$1 AND project=$2 AND valid_to IS NULL FOR UPDATE",
-		rel.SourceID, b.project,
-	).Scan(&dummy); err != nil {
-		if err == pgx.ErrNoRows {
-			return fmt.Errorf("source memory %q does not exist or is invalidated", rel.SourceID)
-		}
-		return fmt.Errorf("check source memory: %w", err)
-	}
-	if err := tx.QueryRow(ctx,
-		"SELECT 1 FROM memories WHERE id=$1 AND project=$2 AND valid_to IS NULL FOR UPDATE",
-		rel.TargetID, b.project,
-	).Scan(&dummy); err != nil {
-		if err == pgx.ErrNoRows {
-			return fmt.Errorf("target memory %q does not exist or is invalidated", rel.TargetID)
-		}
-		return fmt.Errorf("check target memory: %w", err)
-	}
-
-	rel.Project = b.project
-	if _, err := tx.Exec(ctx, `
-		INSERT INTO relationships
-		  (id, source_id, target_id, rel_type, strength, project, created_at)
-		VALUES ($1,$2,$3,$4,$5,$6,$7)
-		ON CONFLICT (source_id, target_id, rel_type)
-		DO UPDATE SET strength = EXCLUDED.strength`,
-		rel.ID, rel.SourceID, rel.TargetID,
-		rel.RelType, rel.Strength, rel.Project, rel.CreatedAt,
-	); err != nil {
-		return err
-	}
-	return tx.Commit(ctx)
-}
-
-// GetConnected performs BFS from memoryID via a single recursive CTE (#113).
-// Returns all connected memories up to maxHops hops away, with the shortest-path
-// metadata (rel_type, direction, strength) for each discovered node.
-func (b *PostgresBackend) GetConnected(ctx context.Context, memoryID string, maxHops int) ([]ConnectedResult, error) {
-	if maxHops <= 0 {
-		return nil, nil
-	}
-	rows, err := b.pool.Query(ctx, `
-WITH RECURSIVE bfs(neighbor_id, rel_type, direction, strength, hop, path) AS (
-  -- Hop 1: direct outgoing neighbors of the seed node.
-  SELECT target_id, rel_type, 'outgoing'::text, strength, 1,
-         ARRAY[$1::text, target_id::text]
-  FROM relationships
-  WHERE source_id = $1 AND project = $2
-  UNION ALL
-  -- Hop 1: direct incoming neighbors of the seed node.
-  SELECT source_id, rel_type, 'incoming'::text, strength, 1,
-         ARRAY[$1::text, source_id::text]
-  FROM relationships
-  WHERE target_id = $1 AND project = $2
-  UNION ALL
-  -- Deeper hops: expand outgoing from current frontier, skip visited nodes.
-  SELECT r.target_id, r.rel_type, 'outgoing'::text, r.strength, b.hop + 1,
-         b.path || r.target_id::text
-  FROM relationships r
-  JOIN bfs b ON b.neighbor_id = r.source_id
-  WHERE r.project = $2 AND b.hop < $3 AND NOT r.target_id = ANY(b.path)
-  UNION ALL
-  -- Deeper hops: expand incoming from current frontier, skip visited nodes.
-  SELECT r.source_id, r.rel_type, 'incoming'::text, r.strength, b.hop + 1,
-         b.path || r.source_id::text
-  FROM relationships r
-  JOIN bfs b ON b.neighbor_id = r.target_id
-  WHERE r.project = $2 AND b.hop < $3 AND NOT r.source_id = ANY(b.path)
-)
-SELECT DISTINCT ON (bfs.neighbor_id)
-  bfs.neighbor_id, bfs.rel_type, bfs.direction, bfs.strength,
-  m.id, m.content, m.memory_type, m.project, m.tags, m.importance,
-  m.access_count, m.last_accessed, m.created_at, m.updated_at,
-  m.immutable, m.expires_at, m.summary, m.content_hash, m.storage_mode,
-  m.search_vector, m.valid_from, m.valid_to, m.invalidation_reason,
-  m.dynamic_importance, m.retrieval_interval_hrs, m.next_review_at,
-  m.times_retrieved, m.times_useful, m.retrieval_precision, m.episode_id
-FROM bfs
-JOIN memories m ON m.id = bfs.neighbor_id AND m.valid_to IS NULL
-ORDER BY bfs.neighbor_id, bfs.hop ASC`,
-		memoryID, b.project, maxHops,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var results []ConnectedResult
-	for rows.Next() {
-		var neighborID, relType, direction string
-		var strength float64
-		var m types.Memory
-		var tagsJSON []byte
-		var searchVector, invalidationReason *string
-		var expiresAt, validTo, nextReviewAt *time.Time
-		err := rows.Scan(
-			&neighborID, &relType, &direction, &strength,
-			&m.ID, &m.Content, &m.MemoryType, &m.Project, &tagsJSON, &m.Importance,
-			&m.AccessCount, &m.LastAccessed, &m.CreatedAt, &m.UpdatedAt,
-			&m.Immutable, &expiresAt, &m.Summary, &m.ContentHash, &m.StorageMode,
-			&searchVector, &m.ValidFrom, &validTo, &invalidationReason,
-			&m.DynamicImportance, &m.RetrievalIntervalHrs, &nextReviewAt,
-			&m.TimesRetrieved, &m.TimesUseful, &m.RetrievalPrecision, &m.EpisodeID,
-		)
-		if err != nil {
-			return nil, err
-		}
-		if tagsJSON != nil {
-			_ = json.Unmarshal(tagsJSON, &m.Tags)
-		}
-		m.ExpiresAt = expiresAt
-		m.ValidTo = validTo
-		m.InvalidationReason = invalidationReason
-		m.NextReviewAt = nextReviewAt
-		results = append(results, ConnectedResult{
-			Memory:    &m,
-			RelType:   relType,
-			Direction: direction,
-			Strength:  strength,
-		})
-	}
-	return results, rows.Err()
-}
-
-func (b *PostgresBackend) BoostEdgesForMemory(ctx context.Context, memoryID string, factor float64) (int, error) {
-	tag, err := b.pool.Exec(ctx, `
-		UPDATE relationships SET strength=LEAST(1.0, strength*$1)
-		WHERE source_id=$2 OR target_id=$2`, factor, memoryID,
-	)
-	return int(tag.RowsAffected()), err
-}
-
-func (b *PostgresBackend) DecayEdgesForMemory(ctx context.Context, memoryID string, factor float64) (int, error) {
-	tag, err := b.pool.Exec(ctx, `
-		UPDATE relationships SET strength=GREATEST(0.0, strength-$1)
-		WHERE source_id=$2 OR target_id=$2`, factor, memoryID,
-	)
-	return int(tag.RowsAffected()), err
-}
-
-func (b *PostgresBackend) GetConnectionCount(ctx context.Context, memoryID, project string) (int, error) {
-	var count int
-	err := b.pool.QueryRow(ctx,
-		"SELECT COUNT(*) FROM relationships WHERE (source_id=$1 OR target_id=$1) AND project=$2",
-		memoryID, project,
-	).Scan(&count)
-	return count, err
-}
-
-// DecayAllEdges decays all edges for a project and prunes those below minStrength.
-// Both operations run inside a single transaction (#109) so a crash between them
-// cannot leave partially-decayed but un-pruned "zombie" edges.
-func (b *PostgresBackend) DecayAllEdges(ctx context.Context, project string, decayFactor, minStrength float64) (int, int, error) {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return 0, 0, err
-	}
-	defer tx.Rollback(ctx)
-
-	tag, err := tx.Exec(ctx, `
-		UPDATE relationships SET strength=GREATEST(0.0, strength-$1)
-		WHERE project=$2`, decayFactor, project,
-	)
-	if err != nil {
-		return 0, 0, err
-	}
-	decayed := int(tag.RowsAffected())
-
-	tag, err = tx.Exec(ctx,
-		"DELETE FROM relationships WHERE strength<$1 AND project=$2",
-		minStrength, project,
-	)
-	if err != nil {
-		return decayed, 0, err
-	}
-	pruned := int(tag.RowsAffected())
-	return decayed, pruned, tx.Commit(ctx)
-}
-
-func (b *PostgresBackend) DeleteRelationshipsForMemory(ctx context.Context, memoryID string) error {
-	_, err := b.pool.Exec(ctx,
-		"DELETE FROM relationships WHERE source_id=$1 OR target_id=$1", memoryID,
-	)
-	return err
-}
-
-func (b *PostgresBackend) GetRelationships(ctx context.Context, project, memoryID string) ([]types.Relationship, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT id, source_id, target_id, rel_type, strength, project, created_at
-		FROM relationships
-		WHERE project = $1 AND (source_id = $2 OR target_id = $2)`,
-		project, memoryID,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var rels []types.Relationship
-	for rows.Next() {
-		var r types.Relationship
-		if err := rows.Scan(&r.ID, &r.SourceID, &r.TargetID, &r.RelType, &r.Strength, &r.Project, &r.CreatedAt); err != nil {
-			return nil, err
-		}
-		rels = append(rels, r)
-	}
-	return rels, rows.Err()
-}
-
-// ── Pruning ───────────────────────────────────────────────────────────────────
-
-// PruneStaleMemories deletes old low-importance and expired memories.
-// Uses RETURNING to emit an audit log line per deleted row (#107) so operators
-// can reconstruct what was pruned from structured logs.
-func (b *PostgresBackend) PruneStaleMemories(ctx context.Context, project string, maxAgeHours float64, maxImportance int) (int, error) {
-	cutoff := time.Now().UTC().Add(-time.Duration(maxAgeHours * float64(time.Hour)))
-	rows, err := b.pool.Query(ctx, `
-		DELETE FROM memories
-		WHERE project=$1 AND NOT immutable AND (
-			(importance>=$2 AND last_accessed<$3 AND access_count=0)
-			OR (expires_at IS NOT NULL AND expires_at<NOW())
-		)
-		RETURNING id, content_hash, importance`, project, maxImportance, cutoff,
-	)
-	if err != nil {
-		return 0, err
-	}
-	defer rows.Close()
-
-	var count int
-	for rows.Next() {
-		var id, contentHash string
-		var importance int
-		if err := rows.Scan(&id, &contentHash, &importance); err != nil {
-			return count, err
-		}
-		slog.Info("prune: deleted stale memory",
-			"project", project, "id", id,
-			"importance", importance, "content_hash", contentHash)
-		count++
-	}
-	return count, rows.Err()
-}
-
-func (b *PostgresBackend) PruneColdDocuments(ctx context.Context, project string, maxAgeHours float64, maxImportance int) (int, error) {
-	cutoff := time.Now().UTC().Add(-time.Duration(maxAgeHours * float64(time.Hour)))
-	tag, err := b.pool.Exec(ctx, `
-		DELETE FROM memories WHERE id IN (
-			SELECT m.id FROM memories m
-			WHERE m.project=$1 AND m.storage_mode='document'
-			  AND NOT m.immutable AND m.importance>=$2 AND m.created_at<$3
-			  AND NOT EXISTS (
-				SELECT 1 FROM chunks c
-				WHERE c.memory_id=m.id AND c.last_matched IS NOT NULL
-			  )
-		)`, project, maxImportance, cutoff,
-	)
-	return int(tag.RowsAffected()), err
-}
-
-// ── Full-text search ──────────────────────────────────────────────────────────
-
-func (b *PostgresBackend) FTSSearch(ctx context.Context, project, query string, limit int, since, before *time.Time) ([]FTSResult, error) {
-	query = strings.TrimSpace(query)
-	if query == "" {
-		return nil, nil
-	}
-
-	q := `SELECT m.*, ts_rank(m.search_vector, plainto_tsquery('english', $1)) AS rank
-		  FROM memories m
-		  WHERE m.search_vector @@ plainto_tsquery('english', $2)
-		  AND m.project=$3 AND m.valid_to IS NULL`
-	args := []any{query, query, project}
-	n := 4
-
-	if since != nil {
-		q += fmt.Sprintf(" AND m.created_at>=$%d", n)
-		args = append(args, since)
-		n++
-	}
-	if before != nil {
-		q += fmt.Sprintf(" AND m.created_at<=$%d", n)
-		args = append(args, before)
-		n++
-	}
-	q += fmt.Sprintf(" ORDER BY rank DESC LIMIT $%d", n)
-	args = append(args, limit)
-
-	rows, err := b.pool.Query(ctx, q, args...)
-	if err != nil {
-		slog.Debug("FTS query failed", "query_len", len(query), "err", err)
-		return nil, fmt.Errorf("FTS search failed: %w", err)
-	}
-	defer rows.Close()
-
-	// Use the shared rowToFTSResult helper to avoid duplicating the 26-column
-	// scan that rowToMemory already owns (#112 — schema drift prevention).
-	return pgx.CollectRows(rows, rowToFTSResult)
-}
-
-func (b *PostgresBackend) RebuildFTS(ctx context.Context) error {
-	// REINDEX CONCURRENTLY must run outside a transaction.
-	conn, err := b.pool.Acquire(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Release()
-	_, err = conn.Exec(ctx, "REINDEX INDEX CONCURRENTLY idx_memories_search")
-	return err
-}
-
-// ── Stats and integrity ───────────────────────────────────────────────────────
-
-func (b *PostgresBackend) GetStats(ctx context.Context, project string) (*types.MemoryStats, error) {
-	stats := &types.MemoryStats{
-		ByType:        map[string]int{},
-		ByImportance:  map[string]int{},
-		Summarization: map[string]any{},
-	}
-
-	if err := b.pool.QueryRow(ctx, "SELECT COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL", project).Scan(&stats.TotalMemories); err != nil {
-		return nil, err
-	}
-	if err := b.pool.QueryRow(ctx, `
-		SELECT COUNT(*) FROM chunks c JOIN memories m ON m.id=c.memory_id WHERE m.project=$1 AND m.valid_to IS NULL`, project,
-	).Scan(&stats.TotalChunks); err != nil {
-		return nil, err
-	}
-	if err := b.pool.QueryRow(ctx, `
-		SELECT COUNT(*) FROM relationships WHERE project=$1`, project,
-	).Scan(&stats.TotalRelationships); err != nil {
-		return nil, err
-	}
-
-	typeRows, err := b.pool.Query(ctx,
-		"SELECT memory_type, COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL GROUP BY memory_type", project)
-	if err != nil {
-		return nil, err
-	}
-	for typeRows.Next() {
-		var mt string
-		var c int
-		if err := typeRows.Scan(&mt, &c); err != nil {
-			typeRows.Close()
-			return nil, err
-		}
-		stats.ByType[mt] = c
-	}
-	typeRows.Close()
-
-	impRows, err := b.pool.Query(ctx,
-		"SELECT importance, COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL GROUP BY importance", project)
-	if err != nil {
-		return nil, err
-	}
-	for impRows.Next() {
-		var imp, c int
-		if err := impRows.Scan(&imp, &c); err != nil {
-			impRows.Close()
-			return nil, err
-		}
-		stats.ByImportance[fmt.Sprintf("%d", imp)] = c
-	}
-	impRows.Close()
-
-	var oldest, newest *time.Time
-	if err := b.pool.QueryRow(ctx, "SELECT MIN(created_at) FROM memories WHERE project=$1 AND valid_to IS NULL", project).Scan(&oldest); err != nil {
-		return nil, err
-	}
-	if err := b.pool.QueryRow(ctx, "SELECT MAX(created_at) FROM memories WHERE project=$1 AND valid_to IS NULL", project).Scan(&newest); err != nil {
-		return nil, err
-	}
-	if oldest != nil {
-		s := oldest.Format(time.RFC3339)
-		stats.Oldest = &s
-	}
-	if newest != nil {
-		s := newest.Format(time.RFC3339)
-		stats.Newest = &s
-	}
-
-	if err := b.pool.QueryRow(ctx,
-		"SELECT COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL AND summary IS NULL", project,
-	).Scan(&stats.PendingSummarization); err != nil {
-		return nil, err
-	}
-
-	var summarized int
-	err = b.pool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM memories WHERE project = $1 AND valid_to IS NULL AND summary IS NOT NULL`,
-		project).Scan(&summarized)
-	if err != nil {
-		summarized = 0
-	}
-	stats.Summarization = map[string]any{
-		"pending":   stats.PendingSummarization,
-		"completed": summarized,
-	}
-
-	if err := b.pool.QueryRow(ctx, "SELECT pg_database_size(current_database())").Scan(&stats.DBSizeBytes); err != nil {
-		return nil, err
-	}
-
-	return stats, nil
-}
-
-func (b *PostgresBackend) ListAllProjects(ctx context.Context) ([]string, error) {
-	rows, err := b.pool.Query(ctx, "SELECT DISTINCT project FROM memories WHERE valid_to IS NULL ORDER BY project")
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var projects []string
-	for rows.Next() {
-		var p string
-		if err := rows.Scan(&p); err != nil {
-			return nil, err
-		}
-		projects = append(projects, p)
-	}
-	return projects, rows.Err()
-}
-
-func (b *PostgresBackend) GetAllMemoryIDs(ctx context.Context, project string) (map[string]struct{}, error) {
-	rows, err := b.pool.Query(ctx, "SELECT id FROM memories WHERE project=$1 AND valid_to IS NULL", project)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	ids := map[string]struct{}{}
-	for rows.Next() {
-		var id string
-		if err := rows.Scan(&id); err != nil {
-			return nil, err
-		}
-		ids[id] = struct{}{}
-	}
-	return ids, rows.Err()
-}
-
-func (b *PostgresBackend) GetMemoriesPendingSummary(ctx context.Context, project string, limit int) ([]IDContent, error) {
-	rows, err := b.pool.Query(ctx,
-		"SELECT id, content FROM memories WHERE project=$1 AND valid_to IS NULL AND summary IS NULL LIMIT $2",
-		project, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var out []IDContent
-	for rows.Next() {
-		var ic IDContent
-		if err := rows.Scan(&ic.ID, &ic.Content); err != nil {
-			return nil, err
-		}
-		out = append(out, ic)
-	}
-	return out, rows.Err()
-}
-
-func (b *PostgresBackend) StoreSummary(ctx context.Context, memoryID, summary string) error {
-	_, err := b.pool.Exec(ctx,
-		"UPDATE memories SET summary=$1 WHERE id=$2 AND project=$3",
-		summary, memoryID, b.project)
-	return err
-}
-
-func (b *PostgresBackend) GetPendingSummaryCount(ctx context.Context, project string) (int, error) {
-	var count int
-	err := b.pool.QueryRow(ctx,
-		"SELECT COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL AND summary IS NULL", project,
-	).Scan(&count)
-	return count, err
-}
-
-func (b *PostgresBackend) GetMemoriesMissingHash(ctx context.Context, project string, limit int) ([]IDContent, error) {
-	rows, err := b.pool.Query(ctx,
-		"SELECT id, content FROM memories WHERE project=$1 AND valid_to IS NULL AND content_hash IS NULL LIMIT $2",
-		project, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var out []IDContent
-	for rows.Next() {
-		var ic IDContent
-		if err := rows.Scan(&ic.ID, &ic.Content); err != nil {
-			return nil, err
-		}
-		out = append(out, ic)
-	}
-	return out, rows.Err()
-}
-
-func (b *PostgresBackend) UpdateMemoryHash(ctx context.Context, memoryID, contentHash string) error {
-	_, err := b.pool.Exec(ctx,
-		"UPDATE memories SET content_hash=$1 WHERE id=$2 AND project=$3",
-		contentHash, memoryID, b.project)
-	return err
-}
-
-func (b *PostgresBackend) GetIntegrityStats(ctx context.Context, project string) (IntegrityStats, error) {
-	var stats IntegrityStats
-	if err := b.pool.QueryRow(ctx, "SELECT COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL", project).Scan(&stats.Total); err != nil {
-		return stats, err
-	}
-	if err := b.pool.QueryRow(ctx,
-		"SELECT COUNT(*) FROM memories WHERE project=$1 AND valid_to IS NULL AND content_hash IS NOT NULL", project,
-	).Scan(&stats.Hashed); err != nil {
-		return stats, err
-	}
-	if err := b.pool.QueryRow(ctx, `
-		SELECT COUNT(*) FROM memories
-		WHERE project=$1 AND valid_to IS NULL AND content_hash IS NOT NULL
-		AND content_hash != encode(sha256(content::bytea),'hex')`, project,
-	).Scan(&stats.Corrupt); err != nil {
-		return stats, err
-	}
-	return stats, nil
-}
-
-// ── Adaptive importance ───────────────────────────────────────────────────────
-
-// UpdateDynamicImportance atomically adjusts dynamic_importance by delta and,
-// when intervalFactor > 0, advances retrieval_interval_hrs *= intervalFactor and
-// sets next_review_at = now + new interval.
-func (b *PostgresBackend) UpdateDynamicImportance(ctx context.Context, id string, delta float64, intervalFactor float64) error {
-	if intervalFactor > 0 {
-		_, err := b.pool.Exec(ctx, `
-			UPDATE memories
-			SET dynamic_importance     = GREATEST(0.1, COALESCE(dynamic_importance, 1.0) + $1),
-			    retrieval_interval_hrs = GREATEST(1, COALESCE(retrieval_interval_hrs, 168) * $2),
-			    next_review_at         = NOW() + (GREATEST(1, COALESCE(retrieval_interval_hrs, 168) * $2) * INTERVAL '1 hour')
-			WHERE id = $3`,
-			delta, intervalFactor, id,
-		)
-		return err
-	}
-	_, err := b.pool.Exec(ctx, `
-		UPDATE memories
-		SET dynamic_importance = GREATEST(0.1, COALESCE(dynamic_importance, 1.0) + $1)
-		WHERE id = $2`,
-		delta, id,
-	)
-	return err
-}
-
-// SetNextReviewAt overrides the next_review_at timestamp for a memory.
-func (b *PostgresBackend) SetNextReviewAt(ctx context.Context, id string, t time.Time) error {
-	_, err := b.pool.Exec(ctx,
-		"UPDATE memories SET next_review_at=$1 WHERE id=$2",
-		t, id,
-	)
-	return err
-}
-
-// DecayStaleImportance multiplies dynamic_importance by factor on all active
-// memories whose next_review_at is in the past. Returns the number of rows updated.
-func (b *PostgresBackend) DecayStaleImportance(ctx context.Context, project string, factor float64) (int, error) {
-	tag, err := b.pool.Exec(ctx, `
-		UPDATE memories
-		SET dynamic_importance = GREATEST(0.1, dynamic_importance * $1),
-		    next_review_at     = next_review_at + (retrieval_interval_hrs * INTERVAL '1 hour')
-		WHERE project = $2
-		  AND valid_to IS NULL
-		  AND next_review_at IS NOT NULL
-		  AND next_review_at < NOW()`,
-		factor, project,
-	)
-	if err != nil {
-		return 0, err
-	}
-	return int(tag.RowsAffected()), nil
-}
-
-// ── Retrieval outcome tracking ────────────────────────────────────────────────
-
-// StoreRetrievalEvent persists a new retrieval event.
-func (b *PostgresBackend) StoreRetrievalEvent(ctx context.Context, event *types.RetrievalEvent) error {
-	resultIDsJSON, err := json.Marshal(event.ResultIDs)
-	if err != nil {
-		return fmt.Errorf("marshal result_ids: %w", err)
-	}
-	_, err = b.pool.Exec(ctx, `
-		INSERT INTO retrieval_events (id, project, query, result_ids, created_at)
-		VALUES ($1, $2, $3, $4, $5)`,
-		event.ID, event.Project, event.Query, resultIDsJSON, event.CreatedAt,
-	)
-	return err
-}
-
-// GetRetrievalEvent fetches a retrieval event by ID. Returns nil, nil if not found.
-func (b *PostgresBackend) GetRetrievalEvent(ctx context.Context, id string) (*types.RetrievalEvent, error) {
-	var event types.RetrievalEvent
-	var resultIDsJSON, feedbackIDsJSON []byte
-	err := b.pool.QueryRow(ctx, `
-		SELECT id, project, query, result_ids, feedback_ids, created_at, feedback_at
-		FROM retrieval_events WHERE id=$1`,
-		id,
-	).Scan(&event.ID, &event.Project, &event.Query, &resultIDsJSON, &feedbackIDsJSON,
-		&event.CreatedAt, &event.FeedbackAt)
-	if err == pgx.ErrNoRows {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-	if len(resultIDsJSON) > 0 {
-		if err := json.Unmarshal(resultIDsJSON, &event.ResultIDs); err != nil {
-			return nil, fmt.Errorf("unmarshal result_ids: %w", err)
-		}
-	}
-	if event.ResultIDs == nil {
-		event.ResultIDs = []string{}
-	}
-	if len(feedbackIDsJSON) > 0 {
-		if err := json.Unmarshal(feedbackIDsJSON, &event.FeedbackIDs); err != nil {
-			return nil, fmt.Errorf("unmarshal feedback_ids: %w", err)
-		}
-	}
-	return &event, nil
-}
-
-// RecordFeedback updates the retrieval event with feedback_ids, sets feedback_at=NOW(),
-// increments times_retrieved on all result memories, increments times_useful on useful
-// memories, and recomputes retrieval_precision once times_retrieved >= 5.
-// RecordFeedback updates the retrieval event with feedback_ids, sets feedback_at=NOW(),
-// increments times_retrieved on all result memories, increments times_useful on useful
-// memories, and recomputes retrieval_precision once times_retrieved >= 5.
-// All four writes are wrapped in a single transaction so a mid-operation crash cannot
-// corrupt the spaced-repetition schedule (fix for #101).
-func (b *PostgresBackend) RecordFeedback(ctx context.Context, eventID string, usefulIDs []string) error {
-	event, err := b.GetRetrievalEvent(ctx, eventID)
-	if err != nil {
-		return fmt.Errorf("RecordFeedback get event: %w", err)
-	}
-	if event == nil {
-		return fmt.Errorf("retrieval event %q not found", eventID)
-	}
-
-	feedbackIDsJSON, err := json.Marshal(usefulIDs)
-	if err != nil {
-		return fmt.Errorf("marshal feedback_ids: %w", err)
-	}
-
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("RecordFeedback begin tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	if _, err := tx.Exec(ctx, `
-		UPDATE retrieval_events SET feedback_ids=$1, feedback_at=NOW() WHERE id=$2`,
-		feedbackIDsJSON, eventID,
-	); err != nil {
-		return fmt.Errorf("update retrieval event: %w", err)
-	}
-
-	// Increment times_retrieved on all result memories.
-	if len(event.ResultIDs) > 0 {
-		if _, err := tx.Exec(ctx, `
-			UPDATE memories SET times_retrieved = COALESCE(times_retrieved, 0) + 1 WHERE id = ANY($1)`,
-			event.ResultIDs,
-		); err != nil {
-			return fmt.Errorf("increment times_retrieved: %w", err)
-		}
-	}
-
-	// Increment times_useful on useful memories.
-	if len(usefulIDs) > 0 {
-		if _, err := tx.Exec(ctx, `
-			UPDATE memories SET times_useful = COALESCE(times_useful, 0) + 1 WHERE id = ANY($1)`,
-			usefulIDs,
-		); err != nil {
-			return fmt.Errorf("increment times_useful: %w", err)
-		}
-	}
-
-	// Recompute precision for result memories that have reached the threshold.
-	if len(event.ResultIDs) > 0 {
-		if _, err := tx.Exec(ctx, `
-			UPDATE memories
-			SET retrieval_precision = CAST(times_useful AS DOUBLE PRECISION) / times_retrieved
-			WHERE id = ANY($1) AND times_retrieved >= 5`,
-			event.ResultIDs,
-		); err != nil {
-			return fmt.Errorf("recompute precision: %w", err)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("RecordFeedback commit: %w", err)
-	}
-	return nil
-}
-
-// IncrementTimesRetrieved increments times_retrieved by 1 on each of the given memory IDs.
-func (b *PostgresBackend) IncrementTimesRetrieved(ctx context.Context, ids []string) error {
-	if len(ids) == 0 {
-		return nil
-	}
-	_, err := b.pool.Exec(ctx, `
-		UPDATE memories SET times_retrieved = COALESCE(times_retrieved, 0) + 1 WHERE id = ANY($1)`,
-		ids,
-	)
-	return err
-}
-
-// ── Episode management ────────────────────────────────────────────────────────
-
-func (b *PostgresBackend) StartEpisode(ctx context.Context, project, description string) (*types.Episode, error) {
-	ep := &types.Episode{
-		ID:          types.NewMemoryID(),
-		Project:     project,
-		Description: description,
-		StartedAt:   time.Now().UTC(),
-	}
-	_, err := b.pool.Exec(ctx, `
-		INSERT INTO episodes (id, project, description, started_at)
-		VALUES ($1, $2, $3, $4)`,
-		ep.ID, ep.Project, ep.Description, ep.StartedAt,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("StartEpisode: %w", err)
-	}
-	return ep, nil
-}
-
-func (b *PostgresBackend) EndEpisode(ctx context.Context, id, summary string) error {
-	_, err := b.pool.Exec(ctx, `
-		UPDATE episodes SET ended_at = NOW(), summary = $1 WHERE id = $2`,
-		summary, id,
-	)
-	return err
-}
-
-func (b *PostgresBackend) ListEpisodes(ctx context.Context, project string, limit int) ([]*types.Episode, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT id, project, description, started_at, COALESCE(ended_at, '0001-01-01'::timestamptz), COALESCE(summary, '')
-		FROM episodes
-		WHERE project = $1
-		ORDER BY started_at DESC
-		LIMIT $2`,
-		project, limit,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("ListEpisodes: %w", err)
-	}
-	defer rows.Close()
-	var eps []*types.Episode
-	for rows.Next() {
-		var ep types.Episode
-		if err := rows.Scan(&ep.ID, &ep.Project, &ep.Description, &ep.StartedAt, &ep.EndedAt, &ep.Summary); err != nil {
-			return nil, err
-		}
-		// Normalise the zero-time sentinel back.
-		if ep.EndedAt.Year() == 1 {
-			ep.EndedAt = time.Time{}
-		}
-		eps = append(eps, &ep)
-	}
-	return eps, rows.Err()
-}
-
-func (b *PostgresBackend) RecallEpisode(ctx context.Context, episodeID string) ([]*types.Memory, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT * FROM memories
-		WHERE episode_id = $1 AND valid_to IS NULL
-		ORDER BY created_at ASC`,
-		episodeID,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("RecallEpisode: %w", err)
-	}
-	return pgx.CollectRows(rows, rowToMemory)
-}
-
-// ── Temporal versioning ───────────────────────────────────────────────────────
-
-// versionMemoryTx snapshots the current state of m into memory_versions.
-// Must be called inside a pgx.Tx that already holds a FOR UPDATE lock on the row.
-func (b *PostgresBackend) versionMemoryTx(ctx context.Context, tx pgx.Tx, m *types.Memory, changeType, changeReason string) error {
-	tagsJSON, err := json.Marshal(m.Tags)
-	if err != nil {
-		return fmt.Errorf("versionMemoryTx: marshal tags: %w", err)
-	}
-	var reason *string
-	if changeReason != "" {
-		reason = &changeReason
-	}
-	now := time.Now().UTC()
-	// Close the system_to window on all prior open versions for this memory.
-	if _, err = tx.Exec(ctx,
-		"UPDATE memory_versions SET system_to=$1 WHERE memory_id=$2 AND system_to IS NULL",
-		now, m.ID,
-	); err != nil {
-		return fmt.Errorf("versionMemoryTx: close prior system_to: %w", err)
-	}
-	_, err = tx.Exec(ctx, `
-		INSERT INTO memory_versions
-			(id, memory_id, content, memory_type, tags, importance,
-			 system_from, valid_from, valid_to, change_type, change_reason, project)
-		VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12)`,
-		types.NewMemoryID(), m.ID, m.Content, m.MemoryType, tagsJSON, m.Importance,
-		now, m.ValidFrom, m.ValidTo, changeType, reason, m.Project,
-	)
-	return err
-}
-
-// GetMemoryHistory returns all version snapshots for memoryID in reverse
-// chronological order (most recent change first).
-func (b *PostgresBackend) GetMemoryHistory(ctx context.Context, project, memoryID string) ([]*types.MemoryVersion, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT id, memory_id, content, memory_type, tags, importance,
-		       system_from, system_to, valid_from, valid_to,
-		       change_type, change_reason, project
-		FROM memory_versions
-		WHERE project=$1 AND memory_id=$2
-		ORDER BY system_from DESC`,
-		project, memoryID,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var out []*types.MemoryVersion
-	for rows.Next() {
-		var v types.MemoryVersion
-		var tagsJSON []byte
-		if err := rows.Scan(
-			&v.ID, &v.MemoryID, &v.Content, &v.MemoryType, &tagsJSON, &v.Importance,
-			&v.SystemFrom, &v.SystemTo, &v.ValidFrom, &v.ValidTo,
-			&v.ChangeType, &v.ChangeReason, &v.Project,
-		); err != nil {
-			return nil, err
-		}
-		if len(tagsJSON) > 0 {
-			if err := json.Unmarshal(tagsJSON, &v.Tags); err != nil {
-				return nil, fmt.Errorf("unmarshal tags: %w", err)
-			}
-		}
-		if v.Tags == nil {
-			v.Tags = []string{}
-		}
-		out = append(out, &v)
-	}
-	return out, rows.Err()
-}
-
-// SoftDeleteMemory marks a memory as invalid by setting valid_to=NOW() and
-// storing the final state in memory_versions with change_type="invalidate".
-// Returns false if not found or already invalidated. Returns error if immutable.
-func (b *PostgresBackend) SoftDeleteMemory(ctx context.Context, project, id, reason string) (bool, error) {
-	tx, err := b.pool.Begin(ctx)
-	if err != nil {
-		return false, err
-	}
-	defer tx.Rollback(ctx) //nolint:errcheck
-
-	row, err := tx.Query(ctx,
-		"SELECT * FROM memories WHERE id=$1 AND project=$2 AND valid_to IS NULL FOR UPDATE",
-		id, project,
-	)
-	if err != nil {
-		return false, err
-	}
-	m, err := pgx.CollectOneRow(row, rowToMemory)
-	if err == pgx.ErrNoRows {
-		return false, nil
-	}
-	if err != nil {
-		return false, err
-	}
-	if m.Immutable {
-		return false, fmt.Errorf("cannot soft-delete immutable memory %s", id)
-	}
-
-	if err := b.versionMemoryTx(ctx, tx, m, types.VersionChangeInvalidate, reason); err != nil {
-		return false, fmt.Errorf("version snapshot: %w", err)
-	}
-
-	var reasonPtr *string
-	if reason != "" {
-		reasonPtr = &reason
-	}
-	now := time.Now().UTC()
-	_, err = tx.Exec(ctx,
-		"UPDATE memories SET valid_to=$1, invalidation_reason=$2 WHERE id=$3 AND project=$4",
-		now, reasonPtr, id, project,
-	)
-	if err != nil {
-		return false, err
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-// GetMemoriesAsOf returns memories that were active at the given point in time:
-// created_at <= asOf AND (valid_to IS NULL OR valid_to > asOf).
-func (b *PostgresBackend) GetMemoriesAsOf(ctx context.Context, project string, asOf time.Time, limit int) ([]*types.Memory, error) {
-	rows, err := b.pool.Query(ctx, `
-		SELECT * FROM memories
-		WHERE project=$1 AND created_at <= $2
-		  AND (valid_to IS NULL OR valid_to > $2)
-		ORDER BY updated_at DESC
-		LIMIT $3`,
-		project, asOf, limit,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	return pgx.CollectRows(rows, rowToMemory)
 }
 
 // ── Row mappers ───────────────────────────────────────────────────────────────
